🔨 Refactor(#339): 신청서 수정페이지 리팩터링
