+++ src/app/(steady)/(application)/application/edit/[steady_id]/[application_id]/page.tsx
@@ -2,15 +2,13 @@
 
 import type { ChangeEvent } from "react";
 import { useState } from "react";
-import { usePathname, useRouter } from "next/navigation";
+import { usePathname } from "next/navigation";
 import { Question, Title } from "@/components/application";
-import { useToast } from "@/components/ui/use-toast";
 import { TextArea } from "@radix-ui/themes";
 import { useSuspenseQuery } from "@tanstack/react-query";
-import editApplication from "@/services/application/editApplication";
 import getApplicationDetails from "@/services/application/getApplicationDetails";
 import getSteadyDetails from "@/services/steady/getSteadyDetails";
-import Button, { buttonSize } from "@/components/_common/Button";
+import EditButtons from "@/components/containers/applicationEdit/EditButtons";
 import {
   getApplicationDetailsKey,
   getSteadyDetailsKey,
@@ -23,15 +21,12 @@ const ApplicationEditPage = ({
 }) => {
   const applicationId = params.application_id;
   const steadyId = params.steady_id;
-  const router = useRouter();
   const pathname = usePathname();
   const pageType = pathname.split("/")[2];
-  const { toast } = useToast();
-  const { data: applicationData, refetch: applicationRefetch } =
-    useSuspenseQuery({
-      queryKey: getApplicationDetailsKey(applicationId),
-      queryFn: () => getApplicationDetails(applicationId),
-    });
+  const { data: applicationData } = useSuspenseQuery({
+    queryKey: getApplicationDetailsKey(applicationId),
+    queryFn: () => getApplicationDetails(applicationId),
+  });
 
   const { data: steadyDetailsData } = useSuspenseQuery({
     queryKey: getSteadyDetailsKey(steadyId),
@@ -45,33 +40,6 @@ const ApplicationEditPage = ({
       answer: survey.answer,
     })),
   );
-
-  const handleClickEdit = async () => {
-    const checkInvalidAnswers = survey.some((item) => !item.answer.length);
-    if (checkInvalidAnswers) {
-      toast({
-        description: "입력되지 않은 항목이 있습니다.",
-        variant: "red",
-      });
-    } else {
-      try {
-        await editApplication(applicationId, {
-          answers: survey.map((item) => item.answer),
-        });
-        toast({
-          description: "스테디 신청서 수정 성공!",
-          variant: "green",
-        });
-        await applicationRefetch();
-        router.replace(`/steady/detail/${steadyId}`);
-      } catch (error) {
-        toast({
-          description: "스테디 신청서 수정 실패!",
-          variant: "red",
-        });
-      }
-    }
-  };
 
   const handleChangeAnswer = (
     event: ChangeEvent<HTMLTextAreaElement>,
@@ -113,18 +81,11 @@ const ApplicationEditPage = ({
         ))}
       </Title>
       <div className="flex items-center justify-end gap-20">
-        <Button
-          className={`${buttonSize.sm} bg-st-white`}
-          onClick={() => router.back()}
-        >
-          취소
-        </Button>
-        <Button
-          onClick={handleClickEdit}
-          className={`${buttonSize.sm} bg-st-primary text-st-white`}
-        >
-          수정 완료
-        </Button>
+        <EditButtons
+          survey={survey}
+          applicationId={applicationId}
+          steadyId={steadyId}
+        />
       </div>
     </>
   );
